pyproject.toml
[project]
name = "dataproduct-mcp"
version = "0.0.1"
description = "MCP server to work with Data Contracts"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.33.0",
    "duckdb>=1.2.2",
    "mcp>=1.1.0",
    "numpy>=2.0.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.30.0",
    "databricks-sdk>=0.1.0",
]

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.389",
]
